requirements.txt
alembic==1.4.3
astroid==2.4.2
autopep8==1.5.4
certifi==2020.12.5
click==7.1.2
elasticsearch==7.10.1
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Migrate==2.5.3
Flask-MySQL==1.5.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.16.0
mccabe==0.6.1
numpy==1.19.4
packaging==21.3
pandas==1.1.5
pika==1.2.1
pycodestyle==2.6.0
PyMySQL==0.10.1
pyparsing==3.0.9
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2020.4
six==1.15.0
SQLAlchemy==1.3.22
toml==0.10.2
typed-ast==1.4.1
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
gunicorn==20.1.0
requests==2.27.1